+++ lib/bootstrap_form/helpers/bootstrap.rb
@@ -1,18 +1,19 @@
 module BootstrapForm
   module Helpers
     module Bootstrap
+
       def button(value = nil, options = {}, &block)
-        options.reverse_merge! class: 'btn btn-secondary'
+        setup_css_class 'btn btn-secondary', options
         super
       end
 
       def submit(name = nil, options = {})
-        options.reverse_merge! class: 'btn btn-secondary'
+        setup_css_class 'btn btn-secondary', options
         super
       end
 
       def primary(name = nil, options = {}, &block)
-        options.reverse_merge! class: 'btn btn-primary'
+        setup_css_class 'btn btn-primary', options
 
         if options[:render_as_button] || block_given?
           options.except! :render_as_button
@@ -66,7 +67,7 @@ module BootstrapForm
           control_class: [options[:control_class], static_class].compact.join(" ")
         })
 
-        static_options[:value] = object.send(name) if static_options[:value].nil?
+        static_options[:value] = object.send(name) unless static_options.has_key?(:value)
 
         text_field_with_bootstrap(name, static_options)
       end
@@ -104,6 +105,19 @@ module BootstrapForm
       def static_class
         "form-control-plaintext"
       end
+
+
+      private
+
+        def setup_css_class(the_class, options = {})
+          unless options.has_key? :class
+            if extra_class = options.delete(:extra_class)
+              the_class = "#{the_class} #{extra_class}"
+            end
+            options[:class] = the_class
+          end
+        end
+
     end
   end
 end