+++ test/bootstrap_other_components_test.rb
@@ -61,6 +61,20 @@ class BootstrapOtherComponentsTest < ActionView::TestCase
     assert_equivalent_xml expected, output
   end
 
+  test "static control support a nil value" do
+    output = @horizontal_builder.static_control label: "Custom Label", value: nil
+
+    expected = <<-HTML.strip_heredoc
+      <div class="form-group row">
+        <label class="col-form-label col-sm-2" for="user_">Custom Label</label>
+        <div class="col-sm-10">
+          <input class="form-control-plaintext" id="user_" name="user[]" readonly="readonly" type="text"/>
+        </div>
+      </div>
+    HTML
+    assert_equivalent_xml expected, output
+  end
+
   test "static control won't overwrite a control_class that is passed by the user" do
     output = @horizontal_builder.static_control :email, control_class: "test_class"
 
@@ -123,6 +137,12 @@ class BootstrapOtherComponentsTest < ActionView::TestCase
                           @builder.button("<span>I'm HTML!</span> in a button!".html_safe)
   end
 
+  test "regular button can have extra css classes" do
+    expected = %{<button class="btn btn-secondary test-button" name="button" type="submit"><span>I'm HTML!</span> in a button!</button>}
+    assert_equivalent_xml expected,
+                          @builder.button("<span>I'm HTML!</span> in a button!".html_safe, extra_class: 'test-button')
+  end
+
   test "submit button defaults to rails action name" do
     expected = %{<input class="btn btn-secondary" name="commit" type="submit" value="Create User" />}
     assert_equivalent_xml expected, @builder.submit
@@ -133,6 +153,11 @@ class BootstrapOtherComponentsTest < ActionView::TestCase
     assert_equivalent_xml expected, @builder.submit("Submit Form")
   end
 
+  test "submit button can have extra css classes" do
+    expected = %{<input class="btn btn-secondary test-button" name="commit" type="submit" value="Submit Form" />}
+    assert_equivalent_xml expected, @builder.submit("Submit Form", extra_class: 'test-button')
+  end
+
   test "override submit button classes" do
     expected = %{<input class="btn btn-primary" name="commit" type="submit" value="Submit Form" />}
     assert_equivalent_xml expected, @builder.submit("Submit Form", class: "btn btn-primary")
@@ -141,6 +166,11 @@ class BootstrapOtherComponentsTest < ActionView::TestCase
   test "primary button uses proper css classes" do
     expected = %{<input class="btn btn-primary" name="commit" type="submit" value="Submit Form" />}
     assert_equivalent_xml expected, @builder.primary("Submit Form")
+  end
+
+  test "primary button can have extra css classes" do
+    expected = %{<input class="btn btn-primary test-button" name="commit" type="submit" value="Submit Form" />}
+    assert_equivalent_xml expected, @builder.primary("Submit Form", extra_class: 'test-button')
   end
 
   test "primary button can render as HTML button" do